requirements.txt
Pillow
discord.py
gTTS
pytesseract
opencv-python

